Move RequestError to WebHDFS scope
