+++ lib/webhdfs/exceptions.rb
@@ -12,3 +12,23 @@ class WebHDFS::ServerError < WebHDFS::Error; end
 class WebHDFS::RequestFailedError < WebHDFS::Error; end
 
 class WebHDFS::KerberosError < WebHDFS::Error; end
+
+class WebHDFS::RequestError < WebHDFS::Error
+  attr_accessor :response, :metadata
+
+  def initialize(response)
+    @response = response
+    @metadata = {
+      headers: response.headers,
+      body: response.body
+    }
+    message = default_message(response)
+    super(message)
+  end
+
+  private
+
+  def default_message(response)
+    "Request failed: #{response.env.method.upcase} #{response.status} #{response.env.url}"
+  end
+end
